Enable workload identiy and split clusters configuration
According to prow announcements,

```
`plank.default_decoration_configs` can optionally be replaced with
`plank.default_decoration_config_entries` which supports a new format
that is a slice of filters with associated decoration configs rather than a
map. Currently entries can filter by repo and/or cluster. The old field is still
supported and will not be deprecated.
```

The configuration is modified to use this field and ensure we can add more build clusters.

- Add annotation to ensure k8s service accounts for prow components can use workload identity.
- bump prow components version to the last version